orb-backend/requirements.txt
django
djangorestframework
python-dotenv
djangorestframework-simplejwt
django-cors-headers

